// src/modules/Nastroenie/ Nastroenie.types.ts
// Собранные уникальные id из всех массивов (NASTROENIE_CHIPS, EMOTIONS_CHIPS, FEELS_CHIPS)
export type TEmotionId =
  | "Super"
  | "Good"
  | "Normal"
  | "Sad"
  | "Terrible"
  | "Melancholy"
  | "Dejection"
  | "Sadness"
  | "Resentment"
  | "Despair"
  | "Frustration"
  | "Fatigue"
  | "Irritation"
  | "Confusion"
  | "Doubt"
  | "Anxiety"
  | "Anger"
  | "Apathy"
  | "Weakness"
  | "Regret"
  | "Boredom"
  | "Rejection";